update group screen WIP
