+++ BrightID/src/components/Groups/IndividualGroup/InviteListScreen.tsx
