+++ BrightID/src/components/Groups/IndividualGroup/MemberCard.tsx
