+++ BrightID/src/components/Groups/old_GroupsScreen.js
